+++ apps/home/package.json
@@ -30,7 +30,6 @@
     "@lightdotso/templates": "workspace:*",
     "@lightdotso/ui": "workspace:*",
     "@lightdotso/utils": "workspace:*",
-    "@lightdotso/wagmi": "workspace:*",
     "@next/bundle-analyzer": "15.0.0-canary.120",
     "@radix-ui/react-icons": "^1.3.0",
     "@sentry/integrations": "^7.114.0",
